Add code for Sequence
This includes a socket, a node, changes to categories, an operator, changes to sdk.
I've decided to dump all nodes in the same category for the time being.
I'm thinking of renaming Body to Geometry and Sequence to Animation.
Also the default values are not the ones I would personally use most often, so I might change those.