Add file_name to Geometry and Animation
This way one can easily get what the file name for the node will be.
It's useful to do this here instead of in the export code, because the QC generator will need it too.
If the object / collection / action isn't set, the name will be Geometry / Animation.
If the suffix isn't set (this is not possible but whatever), it will have no suffix.