Ban reroute nodes, mark invalid links
Reroutes give me too much of a headache, due to various reasons, so I'll just relink and remove them as soon as they're added.
The invalid link check is very simple, just compares types, and I think that's fine.
It gets called from a timer, because I think something gets called after update which resets the link flags, so I have to set them after whatever that is.
Alternatively I could just remove the links as soon as they're made, but I think this is nicer.