Change naming scheme for nodes
Now all the classes in the addon follow the naming scheme of SOURCENODES_class_name.
Whether that's a good thing, I don't know, but at least it's consistent.
Also I changed the icons for the nodes.
And the category is called Model Nodes now, because I'm thinking there might be a place for Material Nodes at some point.